+++ dart/lib/sentry.dart
@@ -3,12 +3,13 @@
 // found in the LICENSE file.
 
 /// A pure Dart client for Sentry.io crash reporting.
+export 'src/default_integrations.dart';
+export 'src/hub.dart';
 export 'src/protocol.dart';
 export 'src/scope.dart';
 export 'src/sentry.dart';
 export 'src/sentry_client.dart';
-export 'src/hub.dart';
 export 'src/sentry_options.dart';
-export 'src/transport/transport.dart';
 // useful for integrations
 export 'src/throwable_mechanism.dart';
+export 'src/transport/transport.dart';
